bazel - initial limited CAS implementation
